pyproject.toml
[tool.poetry]
name = "petstore"
version = "0.1.0"
description = ""
authors = ["anastakhova"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.0"
pytest = "^7.1.0"
pytest-xdist = "^2.5.0"
flake8 = "^5.0.4"
#datamodel-code-generator = "^0.11.19"
swagger-codegen = "^0.1.31"
Werkzeug = "1.0.0"
MarkupSafe = "2.0.1"
allure-pytest = "^2.11.1"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
